+++ packages/terra-core-docs/src/terra-dev-site/doc/paginator/example/PaginatorExample.jsx
@@ -47,12 +47,12 @@ class PaginatorExample extends React.Component {
       <div className={cx('paginator-wrapper')}>
         <Dialog
           header={(
-            <h1>
+            <h1 id="paginator_header_id">
               Page
               {this.state.currentPage}
             </h1>
 )}
-          footer={<Paginator onPageChange={this.changePages} selectedPage={1} totalCount={totalCount} itemCountPerPage={10} />}
+          footer={<Paginator onPageChange={this.changePages} selectedPage={1} totalCount={totalCount} itemCountPerPage={10} ariaLabelledBy="paginator_header_id" />}
         >
           {this.state.content}
         </Dialog>