+++ packages/terra-paginator/src/ControlledProgressivePaginator.jsx
@@ -31,6 +31,14 @@ const propTypes = {
    */
   itemCountPerPage: PropTypes.number,
   /**
+   * Allows user to set the heading id of the page.
+   */
+  ariaLabelledBy: PropTypes.string,
+  /**
+   * Allows user to set the custom paginator label.
+   */
+  ariaLabel: PropTypes.string,
+  /**
    * @private
    * The intl object to be injected for translations.
    */
@@ -78,16 +86,19 @@ class ControlledProgressivePaginator extends React.Component {
       intl,
       totalCount,
       itemCountPerPage,
+      ariaLabelledBy,
+      ariaLabel,
       pageLabel,
     } = this.props;
     const totalPages = (totalCount) ? calculatePages(totalCount, itemCountPerPage) : 0;
     const previousPageIndex = selectedPage === 1 ? 1 : selectedPage - 1;
     const nextPageIndex = selectedPage === totalPages ? totalPages : selectedPage + 1;
+    const paginatorAriaLabel = ariaLabel || 'pagination';
 
     const { messageId, messageAttributes } = getPageLabel(pageLabel, selectedPage, totalPages);
 
-    return (
-      <div className={cx('paginator', 'progressive', theme.className)} role="navigation" aria-label="pagination">
+    const fullViewChildren = (
+      <>
         <div>
           {
             intl.formatMessage({ id: messageId }, messageAttributes)
@@ -145,8 +156,22 @@ class ControlledProgressivePaginator extends React.Component {
             )
           }
         </div>
-      </div>
+      </>
     );
+
+    const fullView = ariaLabelledBy
+      ? (
+        <nav className={cx('paginator', 'progressive', theme.className)} aria-labelledby={ariaLabelledBy}>
+          {fullViewChildren}
+        </nav>
+      )
+      : (
+        <nav className={cx('paginator', 'progressive', theme.className)} aria-label={paginatorAriaLabel}>
+          {fullViewChildren}
+        </nav>
+      );
+
+    return fullView;
   }
 
   reducedProgressivePaginator() {
@@ -156,16 +181,19 @@ class ControlledProgressivePaginator extends React.Component {
       intl,
       totalCount,
       itemCountPerPage,
+      ariaLabelledBy,
+      ariaLabel,
       pageLabel,
     } = this.props;
     const totalPages = (totalCount) ? calculatePages(totalCount, itemCountPerPage) : 0;
     const previousPageIndex = selectedPage === 1 ? 1 : selectedPage - 1;
     const nextPageIndex = selectedPage === totalPages ? totalPages : selectedPage + 1;
+    const paginatorAriaLabel = ariaLabel || 'pagination';
 
     const { messageId, messageAttributes } = getPageLabel(pageLabel, selectedPage, totalPages);
 
-    return (
-      <div className={cx('paginator', theme.className)} role="navigation" aria-label="pagination">
+    const reducedViewChildren = (
+      <>
         <div>
           {
             (totalCount) && (
@@ -225,8 +253,22 @@ class ControlledProgressivePaginator extends React.Component {
             )
           }
         </div>
-      </div>
+      </>
     );
+
+    const reducedView = ariaLabelledBy
+      ? (
+        <nav className={cx('paginator', theme.className)} aria-labelledby={ariaLabelledBy}>
+          {reducedViewChildren}
+        </nav>
+      )
+      : (
+        <nav className={cx('paginator', theme.className)} aria-label={paginatorAriaLabel}>
+          {reducedViewChildren}
+        </nav>
+      );
+
+    return reducedView;
   }
 
   render() {