+++ packages/terra-paginator/tests/jest/ControlledPaginator.test.jsx
@@ -7,6 +7,8 @@ import ControlledPaginator from '../../src/ControlledPaginator';
 
 describe('Paginator', () => {
   const defaultRender = <ControlledPaginator onPageChange={e => typeof e} selectedPage={1} totalCount={2234} itemCountPerPage={20} />;
+  const renderWithHeadingId = <ControlledPaginator onPageChange={e => typeof e} selectedPage={1} totalCount={2234} itemCountPerPage={20} ariaLabelledBy="paginator_header_id" />;
+  const renderWithAriaLabel = <ControlledPaginator onPageChange={e => typeof e} selectedPage={1} totalCount={2234} itemCountPerPage={20} ariaLabel="paginator_label" />;
 
   // Snapshot Tests
   it('should render a Controlled Paginator - Tiny', () => {
@@ -16,7 +18,6 @@ describe('Paginator', () => {
     expect(wrapper).toMatchSnapshot();
   });
 
-  // Snapshot Tests
   it('should render a Controlled Paginator - Large', () => {
     const wrapper = shallowWithIntl(defaultRender).dive();
     wrapper.setState({ breakpoint: 'large' });
@@ -24,13 +25,29 @@ describe('Paginator', () => {
     expect(wrapper).toMatchSnapshot();
   });
 
-  // Snapshot Tests
   it('correctly applies the theme context className', () => {
     const wrapper = mountWithIntl(
       <ThemeContextProvider theme={{ className: 'orion-fusion-theme' }}>
         {defaultRender}
       </ThemeContextProvider>,
     );
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  // Prop Tests
+  it('should render a Controlled Paginator with heading id set - Tiny', () => {
+    const wrapper = shallowWithIntl(renderWithHeadingId).dive();
+    wrapper.setState({ breakpoint: 'tiny' });
+    wrapper.update();
+    expect(wrapper.find('.paginator').prop('aria-labelledby')).toEqual('paginator_header_id');
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('should render a Controlled Paginator with aria label set - Tiny', () => {
+    const wrapper = shallowWithIntl(renderWithAriaLabel).dive();
+    wrapper.setState({ breakpoint: 'tiny' });
+    wrapper.update();
+    expect(wrapper.find('.paginator').prop('aria-label')).toEqual('paginator_label');
     expect(wrapper).toMatchSnapshot();
   });
 });