+++ packages/terra-show-hide/src/index.js
@@ -1,0 +1,6 @@
+import ShowHideFocuser from './ShowHideFocuser';
+import ShowHide from './ShowHide';
+
+export default ShowHide;
+export { ShowHideFocuser };
+