+++ packages/terra-alert/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 4.75.0 - (August 11, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 4.74.0 - (August 11, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 4.73.0 - (August 8, 2023)
 
 * Changed