+++ packages/terra-core-docs/CHANGELOG.md
@@ -5,8 +5,22 @@
 * Added
   * Added accessibility guidance for `terra-arrange`.
   * Added an event to prevent page scroll on hyperlink for `terra-toggle-section-header`.
+  * Added Search field examples for optional Label with default and custom labels.
 
 * Changed
+  * Updated default search delay to 2500ms.
+  * Update Search field examples to be more functionality focused. 
+
+## 1.36.0 - (August 11, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 1.35.0 - (August 11, 2023)
+
+* Changed
+  * Updated all paginator examples for `terra-paginator` to use new props.
+  * Updated imports from `src` to `lib`.
   * Updated `terra-toggle-button` example and user documentation.
 
 ## 1.34.0 - (August 8, 2023)
@@ -32,6 +46,7 @@
 * Updated
   * Updated an example for `terra-dropdown-button`.
   * Updated email field validation for `terra-form-field`.
+  * Updated `terra-show-hide` examples for new focusRef prop, description for the new prop usage.
 
 * Added
   * Added documentation updates for `terra-form-input`.